voc_table/backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
alembic>=1.12.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.1.4
openpyxl>=3.1.2
pymysql>=1.1.0
cryptography>=41.0.0
PyJWT>=2.8.0
llama-cpp-python>=0.2.11
email-validator>=2.1.0
